# frontend/openapi.json
{
  "components": {
    "schemas": {
      "ActivationSteeringExperimentRequest": {
        "additionalProperties": false,
        "properties": {
          "input": {
            "$ref": "#/components/schemas/ActivationSteeringRequest"
          },
          "technique_id": {
            "const": "activation_steering",
            "title": "Technique Id",
            "type": "string"
          }
        },
        "required": [
          "technique_id",
          "input"
        ],
        "title": "ActivationSteeringExperimentRequest",
        "type": "object"
      },
...
      "ActivationSteeringRequest": {
        "additionalProperties": false,
        "properties": {
          "layer": {
            "default": 18,
            "maximum": 128.0,
            "minimum": 0.0,
            "title": "Layer",
            "type": "integer"
          },
          "max_new_tokens": {
            "default": 96,
            "maximum": 128.0,
            "minimum": 1.0,
            "title": "Max New Tokens",
            "type": "integer"
          },
          "model_key": {
            "default": "qwen3-1.7b",
            "title": "Model Key",
            "type": "string"
          },
          "negative_examples": {
            "items": {
              "type": "string"
            },
            "maxItems": 8,
            "minItems": 1,
            "title": "Negative Examples",
            "type": "array"
          },
          "positive_examples": {
            "items": {
              "type": "string"
            },
            "maxItems": 8,
            "minItems": 1,
            "title": "Positive Examples",
            "type": "array"
          },
          "prompt": {
            "maxLength": 4000,
            "minLength": 1,
            "title": "Prompt",
            "type": "string"
          },
          "seed": {
            "default": 16,
            "maximum": 2147483647.0,
            "minimum": 0.0,
            "title": "Seed",
            "type": "integer"
          },
          "strength": {
            "default": 1.0,
            "maximum": 100.0,
            "minimum": -100.0,
            "title": "Strength",
            "type": "number"
          },
          "temperature": {
            "default": 0.7,
            "maximum": 2.0,
            "minimum": 0.0,
            "title": "Temperature",
            "type": "number"
          },
          "top_p": {
            "default": 0.9,
            "exclusiveMinimum": 0.0,
            "maximum": 1.0,
            "title": "Top P",
            "type": "number"
          }
        },
        "required": [
          "prompt",
          "positive_examples",
          "negative_examples"
        ],
        "title": "ActivationSteeringRequest",
        "type": "object"
      },
      "ActivationSteeringResponse": {
        "properties": {
          "baseline_message": {
            "title": "Baseline Message",
            "type": "string"
          },
          "direction_norm": {
            "title": "Direction Norm",
            "type": "number"
          },
          "metadata": {
            "$ref": "#/components/schemas/ActivationSteeringMetadata"
          },
          "model_key": {
            "title": "Model Key",
            "type": "string"
          },
          "prompt": {
            "title": "Prompt",
            "type": "string"
          },
          "steered_message": {
            "title": "Steered Message",
            "type": "string"
          },
          "warnings": {
            "items": {
              "type": "string"
            },
            "title": "Warnings",
            "type": "array"
          }
        },
        "required": [
          "model_key",
          "prompt",
          "baseline_message",
          "steered_message",
          "direction_norm",
          "metadata"
        ],
        "title": "ActivationSteeringResponse",
        "type": "object"
      },
      "ExperimentEnvelope": {
        "properties": {
          "experiment_id": {
            "title": "Experiment Id",
            "type": "string"
          },
          "finished_at": {
            "format": "date-time",
            "title": "Finished At",
            "type": "string"
          },
          "lineage_operation": {
            "anyOf": [
              {
                "enum": [
                  "replay",
                  "fork"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Lineage Operation"
          },
          "parent_experiment_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parent Experiment Id"
          },
          "request": {
            "discriminator": {
              "mapping": {
                "activation_steering": "#/components/schemas/ActivationSteeringExperimentRequest",
                "jacobian_lens": "#/components/schemas/JacobianLensExperimentRequest"
              },
              "propertyName": "technique_id"
            },
            "oneOf": [
              {
                "$ref": "#/components/schemas/JacobianLensExperimentRequest"
              },
              {
                "$ref": "#/components/schemas/ActivationSteeringExperimentRequest"
              }
            ],
            "title": "Request"
          },
          "result": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/JacobianLensResponse"
              },
              {
                "$ref": "#/components/schemas/ActivationSteeringResponse"
              }
            ],
            "title": "Result"
          },
          "schema_version": {
            "const": 1,
            "default": 1,
            "title": "Schema Version",
            "type": "integer"
          },
          "started_at": {
            "format": "date-time",
            "title": "Started At",
            "type": "string"
          },
          "status": {
            "const": "complete",
            "default": "complete",
            "title": "Status",
            "type": "string"
          },
          "technique_id": {
            "title": "Technique Id",
            "type": "string"
          }
        },
        "required": [
          "experiment_id",
          "technique_id",
          "started_at",
          "finished_at",
          "request",
          "result"
        ],
        "title": "ExperimentEnvelope",
        "type": "object"
      },
      "ExperimentForkRequest": {
        "additionalProperties": false,
        "properties": {
          "request": {
            "discriminator": {
              "mapping": {
                "activation_steering": "#/components/schemas/ActivationSteeringExperimentRequest",
                "jacobian_lens": "#/components/schemas/JacobianLensExperimentRequest"
              },
              "propertyName": "technique_id"
            },
            "oneOf": [
              {
                "$ref": "#/components/schemas/JacobianLensExperimentRequest"
              },
              {
                "$ref": "#/components/schemas/ActivationSteeringExperimentRequest"
              }
            ],
            "title": "Request"
          }
        },
        "required": [
          "request"
        ],
        "title": "ExperimentForkRequest",
        "type": "object"
      },
      "ExperimentList": {
        "properties": {
          "experiments": {
            "items": {
              "$ref": "#/components/schemas/ExperimentEnvelope"
            },
            "title": "Experiments",
            "type": "array"
          }
        },
        "required": [
          "experiments"
        ],
        "title": "ExperimentList",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "InputToken": {
        "properties": {
          "position": {
            "title": "Position",
            "type": "integer"
          },
          "text": {
            "title": "Text",
            "type": "string"
          },
          "token_id": {
            "title": "Token Id",
            "type": "integer"
          }
        },
        "required": [
          "position",
          "token_id",
          "text"
        ],
        "title": "InputToken",
        "type": "object"
      },
      "JacobianLensExperimentRequest": {
        "additionalProperties": false,
        "properties": {
          "input": {
            "$ref": "#/components/schemas/JacobianLensRequest"
          },
          "technique_id": {
            "const": "jacobian_lens",
            "title": "Technique Id",
            "type": "string"
          }
        },
        "required": [
          "technique_id",
          "input"
        ],
        "title": "JacobianLensExperimentRequest",
        "type": "object"
      },
      "JacobianLensMetadata": {
        "properties": {
          "cache": {
            "const": "modal_volume",
            "default": "modal_volume",
            "title": "Cache",
            "type": "string"
          },
          "elapsed_ms": {
            "title": "Elapsed Ms",
            "type": "integer"
          },
          "jlens_revision": {
            "title": "Jlens Revision",
            "type": "string"
          },
          "lens_file": {
            "title": "Lens File",
            "type": "string"
          },
          "lens_repo": {
            "title": "Lens Repo",
            "type": "string"
          },
          "lens_revision": {
            "title": "Lens Revision",
            "type": "string"
          },
          "max_tokens": {
            "title": "Max Tokens",
            "type": "integer"
          },
          "model_id": {
            "title": "Model Id",
            "type": "string"
          },
          "model_revision": {
            "title": "Model Revision",
            "type": "string"
          },
          "source_layers": {
            "items": {
              "type": "integer"
            },
            "title": "Source Layers",
            "type": "array"
          },
          "top_k": {
            "title": "Top K",
            "type": "integer"
          },
          "vocab_size": {
            "title": "Vocab Size",
            "type": "integer"
          }
        },
        "required": [
          "model_id",
          "model_revision",
          "lens_repo",
          "lens_revision",
          "lens_file",
          "jlens_revision",
          "max_tokens",
          "top_k",
          "source_layers",
          "elapsed_ms",
          "vocab_size"
        ],
        "title": "JacobianLensMetadata",
        "type": "object"
      },
      "JacobianLensRequest": {
        "additionalProperties": false,
        "properties": {
          "max_tokens": {
            "default": 64,
            "maximum": 128.0,
            "minimum": 1.0,
            "title": "Max Tokens",
            "type": "integer"
          },
          "model_key": {
            "default": "qwen3-1.7b",
            "title": "Model Key",
            "type": "string"
          },
          "prompt": {
            "maxLength": 4000,
            "minLength": 1,
            "title": "Prompt",
            "type": "string"
          },
          "top_k": {
            "default": 10,
            "maximum": 10.0,
            "minimum": 1.0,
            "title": "Top K",
            "type": "integer"
          }
        },
        "required": [
          "prompt"
        ],
        "title": "JacobianLensRequest",
        "type": "object"
      },
      "JacobianLensResponse": {
        "properties": {
          "metadata": {
            "$ref": "#/components/schemas/JacobianLensMetadata"
          },
          "model_key": {
            "title": "Model Key",
            "type": "string"
          },
          "prompt": {
            "title": "Prompt",
            "type": "string"
          },
          "rank_tracks": {
            "items": {
              "$ref": "#/components/schemas/RankTrack"
            },
            "title": "Rank Tracks",
            "type": "array"
          },
          "rows": {
            "items": {
              "$ref": "#/components/schemas/LayerReadout"
            },
            "title": "Rows",
            "type": "array"
          },
          "tokens": {
            "items": {
              "$ref": "#/components/schemas/InputToken"
            },
            "title": "Tokens",
            "type": "array"
          }
        },
        "required": [
          "model_key",
          "prompt",
          "tokens",
          "rows",
          "metadata"
        ],
        "title": "JacobianLensResponse",
        "type": "object"
      },
      "LayerReadout": {
        "properties": {
          "kind": {
            "enum": [
              "jacobian_lens",
              "model_output"
            ],
            "title": "Kind",
            "type": "string"
          },
          "layer": {
            "title": "Layer",
            "type": "integer"
          },
          "positions": {
            "items": {
              "$ref": "#/components/schemas/PositionReadout"
            },
            "title": "Positions",
            "type": "array"
          }
        },
        "required": [
          "layer",
          "kind",
          "positions"
        ],
        "title": "LayerReadout",
        "type": "object"
      },
...
      "PublicError": {
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "retryable": {
            "default": false,
            "title": "Retryable",
            "type": "boolean"
          }
        },
        "required": [
          "code",
          "message"
        ],
        "title": "PublicError",
        "type": "object"
      },
      "RankTrack": {
        "properties": {
          "ranks": {
            "items": {
              "items": {
                "type": "integer"
              },
              "type": "array"
            },
            "title": "Ranks",
            "type": "array"
          },
          "text": {
            "title": "Text",
            "type": "string"
          },
          "token_id": {
            "title": "Token Id",
            "type": "integer"
          }
        },
        "required": [
          "token_id",
          "text",
          "ranks"
        ],
        "title": "RankTrack",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "Capability catalog and experiment gateway for Mechanoscope.",
    "title": "Mechanoscope API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/experiments": {
      "get": {
        "operationId": "list_experiments_api_experiments_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 25,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExperimentList"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Experiments",
        "tags": [
          "experiments"
        ]
      }
    },
    "/api/experiments/run": {
      "post": {
        "operationId": "run_experiment_api_experiments_run_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "discriminator": {
                  "mapping": {
                    "activation_steering": "#/components/schemas/ActivationSteeringExperimentRequest",
                    "jacobian_lens": "#/components/schemas/JacobianLensExperimentRequest"
                  },
                  "propertyName": "technique_id"
                },
                "oneOf": [
                  {
                    "$ref": "#/components/schemas/JacobianLensExperimentRequest"
                  },
                  {
                    "$ref": "#/components/schemas/ActivationSteeringExperimentRequest"
                  }
                ],
                "title": "Request"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExperimentEnvelope"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicError"
                }
              }
            },
            "description": "Service Unavailable"
          },
          "504": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicError"
                }
              }
            },
            "description": "Gateway Timeout"
          }
        },
        "summary": "Run Experiment",
        "tags": [
          "experiments"
        ]
      }
    },
    "/api/experiments/{experiment_id}": {
      "delete": {
        "operationId": "delete_experiment_api_experiments__experiment_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "experiment_id",
            "required": true,
            "schema": {
              "title": "Experiment Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Experiment",
        "tags": [
          "experiments"
        ]
      },
      "get": {
        "operationId": "get_experiment_api_experiments__experiment_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "experiment_id",
            "required": true,
            "schema": {
              "title": "Experiment Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExperimentEnvelope"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Experiment",
        "tags": [
          "experiments"
        ]
      }
    },
    "/api/experiments/{experiment_id}/fork": {
      "post": {
        "operationId": "fork_experiment_api_experiments__experiment_id__fork_post",
        "parameters": [
          {
            "in": "path",
            "name": "experiment_id",
            "required": true,
            "schema": {
              "title": "Experiment Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExperimentForkRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExperimentEnvelope"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Fork Experiment",
        "tags": [
          "experiments"
        ]
      }
    },
    "/api/experiments/{experiment_id}/replay": {
      "post": {
        "operationId": "replay_experiment_api_experiments__experiment_id__replay_post",
        "parameters": [
          {
            "in": "path",
            "name": "experiment_id",
            "required": true,
            "schema": {
              "title": "Experiment Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExperimentEnvelope"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Replay Experiment",
        "tags": [
          "experiments"
        ]
      }
    },
    "/api/jlens/run": {
      "post": {
        "operationId": "run_jacobian_lens_api_jlens_run_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/JacobianLensRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JacobianLensResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicError"
                }
              }
            },
            "description": "Service Unavailable"
          },
          "504": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicError"
                }
              }
            },
            "description": "Gateway Timeout"
          }
        },
        "summary": "Run Jacobian Lens",
        "tags": [
          "jacobian-lens"
        ]
      }
    },
...
    "/api/steer": {
      "post": {
        "operationId": "run_activation_steering_api_steer_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ActivationSteeringRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActivationSteeringResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicError"
                }
              }
            },
            "description": "Service Unavailable"
          },
          "504": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicError"
                }
              }
            },
            "description": "Gateway Timeout"
          }
        },
        "summary": "Run Activation Steering",
        "tags": [
          "activation-steering"
        ]
      }
    },
...
  }
}